Permission and quota check for EC2 api.

+++ clc/modules/cluster-manager/src/main/java/com/eucalyptus/network/NetworkGroupManager.java
@@ -74,6 +74,15 @@ public class NetworkGroupManager {
   
   public CreateSecurityGroupResponseType create( CreateSecurityGroupType request ) throws EucalyptusCloudException {
     Context ctx = Contexts.lookup();
+    String action = PolicySpec.requestToAction( request );
+    if ( !ctx.hasAdministrativePrivileges( ) ) {
+      if ( !Permissions.isAuthorized( PolicySpec.EC2_RESOURCE_SECURITYGROUP, "", ctx.getAccount( ), action, ctx.getUser( ) ) ) {
+        throw new EucalyptusCloudException( "Not authorized to create network group for " + ctx.getUser( ) );
+      }
+      if ( !Permissions.canAllocate( PolicySpec.EC2_RESOURCE_SECURITYGROUP, "", action, ctx.getUser( ), 1 ) ) {
+        throw new EucalyptusCloudException( "Quota exceeded to create network group for " + ctx.getUser( ) );
+      }
+    }
     NetworkGroupUtil.makeDefault( ctx.getUserFullName( ) );//ensure the default group exists to cover some old broken installs
     CreateSecurityGroupResponseType reply = ( CreateSecurityGroupResponseType ) request.getReply( );
     NetworkRulesGroup newGroup = new NetworkRulesGroup( ctx.getUserFullName( ), request.getGroupName( ), request.getGroupDescription( ) ); 
@@ -105,14 +114,17 @@ public class NetworkGroupManager {
         throw new EucalyptusCloudException( "Deleting security failed because of: " + ex.getMessage( ) + " for request " + request.toSimpleString( ) );
       }
     } else {
+      if ( !Permissions.isAuthorized( PolicySpec.EC2_RESOURCE_SECURITYGROUP, request.getGroupName( ), ctx.getAccount( ), PolicySpec.requestToAction( request ), ctx.getUser( ) ) ) {
+        throw new EucalyptusCloudException( "Not authorized to delete network group " + request.getGroupName( ) + " for " + ctx.getUser( ) );
+      }
       NetworkGroupUtil.deleteUserNetworkRulesGroup( ctx.getUserFullName( ), request.getGroupName( ) );
     }
     reply.set_return( true );
     return reply;
   }
   
-  public DescribeSecurityGroupsResponseType describe( DescribeSecurityGroupsType request ) throws EucalyptusCloudException {
-    Context ctx = Contexts.lookup();
+  public DescribeSecurityGroupsResponseType describe( final DescribeSecurityGroupsType request ) throws EucalyptusCloudException {
+    final Context ctx = Contexts.lookup();
     NetworkGroupUtil.makeDefault( ctx.getUserFullName( ) );//ensure the default group exists to cover some old broken installs
     final List<String> groupNames = request.getSecurityGroupSet( );
     DescribeSecurityGroupsResponseType reply = ( DescribeSecurityGroupsResponseType ) request.getReply( );
@@ -137,6 +149,9 @@ public class NetworkGroupManager {
                                                               new Predicate<SecurityGroupItemType>( ) {
                                                                 @Override
                                                                 public boolean apply( SecurityGroupItemType arg0 ) {
+                                                                  if ( Permissions.isAuthorized( PolicySpec.EC2_RESOURCE_SECURITYGROUP, arg0.getGroupName( ), ctx.getAccount( ), PolicySpec.requestToAction( request ), ctx.getUser( ) ) ) {
+                                                                    return false;
+                                                                  }
                                                                   return groupNames.isEmpty( ) || groupNames.contains( arg0.getGroupName( ) );
                                                                 }
                                                               } ) ) {
@@ -154,6 +169,9 @@ public class NetworkGroupManager {
     NetworkGroupUtil.makeDefault( ctx.getUserFullName( ) );//ensure the default group exists to cover some old broken installs
     RevokeSecurityGroupIngressResponseType reply = ( RevokeSecurityGroupIngressResponseType ) request.getReply( );
     NetworkRulesGroup ruleGroup = NetworkGroupUtil.getUserNetworkRulesGroup( ctx.getUserFullName( ), request.getGroupName( ) );
+    if ( !ctx.hasAdministrativePrivileges( ) && !Permissions.isAuthorized( PolicySpec.EC2_RESOURCE_SECURITYGROUP, request.getGroupName( ), ctx.getAccount( ), PolicySpec.requestToAction( request ), ctx.getUser( ) ) ) {
+      throw new EucalyptusCloudException( "Not authorized to revoke network group " + request.getGroupName( ) + " for " + ctx.getUser( ) );
+    }
     final List<NetworkRule> ruleList = Lists.newArrayList( );
     for ( IpPermissionType ipPerm : request.getIpPermissions( ) ) {
       ruleList.addAll( NetworkGroupUtil.getNetworkRules( ipPerm ) );
@@ -207,6 +225,9 @@ public class NetworkGroupManager {
     AuthorizeSecurityGroupIngressResponseType reply = ( AuthorizeSecurityGroupIngressResponseType ) request.getReply( );
     EntityWrapper<NetworkRulesGroup> db = NetworkGroupUtil.getEntityWrapper( );
     NetworkRulesGroup ruleGroup = NetworkGroupUtil.getUserNetworkRulesGroup( ctx.getUserFullName( ), request.getGroupName( ) );
+    if ( !ctx.hasAdministrativePrivileges( ) && !Permissions.isAuthorized( PolicySpec.EC2_RESOURCE_SECURITYGROUP, request.getGroupName( ), ctx.getAccount( ), PolicySpec.requestToAction( request ), ctx.getUser( ) ) ) {
+      throw new EucalyptusCloudException( "Not authorized to authorize network group " + request.getGroupName( ) + " for " + ctx.getUser( ) );
+    }
     final List<NetworkRule> ruleList = Lists.newArrayList( );
     for ( IpPermissionType ipPerm : request.getIpPermissions( ) ) {
       try {
